Refactor: Move Homepage to root of pages directory
